backend/requirements.txt
django>=6.0
djangorestframework>=3.17
django-cors-headers>=4.9
djangorestframework-simplejwt>=5.5
pillow>=12.0
python-dotenv>=1.2
requests>=2.32
youtube-transcript-api>=0.6.2
PyPDF2>=3.0.0
whitenoise>=6.7.0
django-unfold>=0.40.0
edge-tts>=6.1.13
